accession2gi {primerTree}	R Documentation

Maps a genbank accession to a nuclotide database gi.

Description

Maps a genbank accession to a nuclotide database gi.

Usage

accession2gi(accession)

Arguments

accession
accession character vector to lookup.

Value

named vector of gis.

[Package primerTree version 1.0.3 Index]