calc_rank_dist_ave {primerTree}	R Documentation

calc_rank_dist_ave Summarize pairwise differences.

Description

calc_rank_dist_ave Summarize pairwise differences.

Usage

calc_rank_dist_ave(x, ranks = common_ranks, ...)

Arguments

x
a primerTree object

ranks
ranks to show unique counts for, defaults to the common ranks

...
Ignored options

Details

The purpose of this function is to calculate the average number of nucleotide differences between species within a given taxonomic level.

There are several key assumptions and calculations made in this function.

First, the function randomly selects one sequence from each species in the primerTree results. This is to keep any one species (e.g. human, cow, etc.) with many hits from skewing the results.

Second, for each taxonomic level tested, the function divides the sequences by each taxon at that level and calculates the mean number of nucleotide differences within that taxa, then returns the mean of those values.

Third, when calculating the average distance, any taxa for which there is only one species is omitted, as the number of nucleotide differences will always be 0.

Value

returns a data frame of results

Examples

## Not run: 
calc_rank_dist_ave(mammals_16S)

calc_rank_dist_ave(bryophytes_trnL)

# Note that the differences between the results from these two primers
# the mean nucleotide differences is much higher for the mammal primers
# than the byrophyte primers. This suggests that the mammal primers have
# better resolution to distinguish individual species.

## End(Not run)

[Package primerTree version 1.0.3 Index]