clustalo {primerTree}	R Documentation

Multiple sequence alignment with clustal omega

Description

Calls clustal omega to align a set of sequences of class DNAbin. Run without any arguments to see all the options you can pass to the command line clustal omega.

Usage

clustalo(x, exec = "clustalo", quiet = TRUE, original.ordering = TRUE,
  ...)

Arguments

x
an object of class 'DNAbin'

exec
a character string with the name or path to the program

quiet
whether to supress output to stderr or stdout

original.ordering
use the original ordering of the sequences

...
additional arguments passed to the command line clustalo

[Package primerTree version 1.0.3 Index]