filter_seqs {primerTree}	R Documentation

Filter out sequences retrieved by search_primer_pair() that are either too short or too long. The alignment and tree will be recalculated after removing unwanted reads.

Description

Filter out sequences retrieved by search_primer_pair() that are either too short or too long. The alignment and tree will be recalculated after removing unwanted reads.

Usage

filter_seqs(x, ...)

## S3 method for class 'primerTree'
filter_seqs(x, min_length = 0, max_length = Inf, ...)

Arguments

x
a primerTree object

...
additional arguments passed to methods.

min_length
the minimum sequence length to keep

max_length
the maximum sequence length to keep

Value

a primerTree object

Methods (by class)

primerTree: Method for primerTree objects

Examples

## Not run: 
# filter out sequences longer or shorter than desired:
mammals_16S_filtered <- filter_seqs(mammals_16S, min_length=131, max_length=156)

## End(Not run)

[Package primerTree version 1.0.3 Index]