get_sequences {primerTree}	R Documentation

Retrieves fasta sequences from NCBI nucleotide database.

Description

Retrieves fasta sequences from NCBI nucleotide database.

Usage

get_sequences(gi, start = NULL, stop = NULL, simplify = TRUE,
  .parallel = FALSE, .progress = "none")

Arguments

gi
the gi number of the sequence to retrieve

start
start bases to retrieve, numbered beginning at 1. If NULL the beginning of the sequence.

stop
stop bases to retrieve, numbered beginning at 1. if NULL the stop of the sequence.

simplify
simplify the FASTA headers to include only the genbank accession.

.parallel
if 'TRUE', perform in parallel, using parallel backend provided by foreach

.progress
name of the progress bar to use, see 'create_progress_bar'

Value

an DNAbin object.

See Also

DNAbin

[Package primerTree version 1.0.3 Index]