gi2taxid {primerTree}	R Documentation

Maps a nucleotide database gi to a taxonomy database taxId

Description

Maps a nucleotide database gi to a taxonomy database taxId

Usage

gi2taxid(gi)

Arguments

gi
gi character vector to lookup.

Value

named vector of taxIds.

[Package primerTree version 1.0.3 Index]