identify.primerTree_plot {primerTree}	R Documentation

identify the point closest to the mouse click only works on single ranks

Description

identify the point closest to the mouse click only works on single ranks

Usage

## S3 method for class 'primerTree_plot'
identify(x, ...)

Arguments

x
the plot to identify

...
additional arguments passed to annotate

[Package primerTree version 1.0.3 Index]