layout_tree_ape {primerTree}	R Documentation

layout a tree using ape, return an object to be plotted by plot_tree

Description

layout a tree using ape, return an object to be plotted by plot_tree

Usage

layout_tree_ape(tree, ...)

Arguments

tree
The phylo tree to be plotted

...
additional arguments to plot.phylo

Value

edge
list of x, y and xend, yend coordinates as well as ids for the edges

tips
list of x, y, label and id for the tips

nodes
list of x, y and id for the nodes

[Package primerTree version 1.0.3 Index]