plot_tree {primerTree}	R Documentation

plots a tree, optionally with colored and labeled points by taxonomic rank

Description

plots a tree, optionally with colored and labeled points by taxonomic rank

Usage

plot_tree(tree, type = "unrooted", main = NULL, guide_size = NULL,
  rank = NULL, taxonomy = NULL, size = 2, legend_cutoff = 25, ...)

Arguments

tree
to be plotted, use layout_tree to layout tree.

type
The type of tree to plot, default unrooted.

main
An optional title for the plot

guide_size
The size of the length guide. If NULL auto detects a reasonable size.

rank
The rank to include, if null only the tree is plotted

taxonomy
A data.frame with an accession field corresponding to the tree tip labels.

size
The size of the colored points

legend_cutoff
The number of different taxa names after which the names are no longer printed.

...
additional arguments passed to layout_tree_ape

Value

plot to be printed.

[Package primerTree version 1.0.3 Index]