primerTree {primerTree}	R Documentation

primerTree Visually Assessing the Specificity and Informativeness of Primer Pairs

Description

primerTree has two main commands: search_primer_pair which takes a primer pair and returns an primerTree object of the search results plot.primerTree a S3 method for plotting the primerTree object obtained using search_primer_pair

[Package primerTree version 1.0.3 Index]