primer_search {primerTree}	R Documentation

Query a pair of primers using ncbi's Primer-BLAST, if primers contain iupac

Description

ambiguity codes, enumerate all possible combinations and combine the results.

Usage

primer_search(forward, reverse, num_aligns = 500, num_permutations = 25,
  ..., .parallel = FALSE, .progress = "none")

Arguments

forward
forward primer to search by 5'-3' on plus strand

reverse
reverse primer to search by 5'-3' on minus strand

num_aligns
number of alignment results to keep

num_permutations
the number of primer permutations to search, if the degenerate bases cause more than this number of permutations to exist, this number will be sampled from all possible permutations.

...
additional arguments passed to Primer-Blast

.parallel
if 'TRUE', perform in parallel, using parallel backend provided by foreach

.progress
name of the progress bar to use, see 'create_progress_bar'

Value

data.frame of primer hits

[Package primerTree version 1.0.3 Index]