search_primer_pair {primerTree}	R Documentation

Automatic primer searching Search a given primer pair, retrieving the alignment results, their product sequences, the taxonomic information for the sequences, a multiple alignment of the products

Description

Automatic primer searching Search a given primer pair, retrieving the alignment results, their product sequences, the taxonomic information for the sequences, a multiple alignment of the products

Usage

search_primer_pair(forward, reverse, name = NULL, num_aligns = 500,
  num_permutations = 25, simplify = TRUE, clustal_options = list(),
  distance_options = list(model = "N", pairwise.deletion = T), ...,
  .parallel = FALSE, .progress = "none")

Arguments

forward
forward primer to search by 5'-3' on plus strand

reverse
reverse primer to search by 5'-3' on minus strand

name
name to give to the primer pair

num_aligns
number of alignment results to keep

num_permutations
the number of primer permutations to search, if the degenerate bases cause more than this number of permutations to exist, this number will be sampled from all possible permutations.

simplify
use simple names for primer hit results or complex

clustal_options
a list of options to pass to clustal omega, see link{clustalo} for a list of options

distance_options
a list of options to pass to dist.dna, see link{dist.dna} for a list of options

...
additional arguments passed to Primer-Blast

.parallel
if 'TRUE', perform in parallel, using parallel backend provided by foreach

.progress
name of the progress bar to use, see create_progress_bar

Value

A list with the following elements,

name
name of the primer pair

BLAST_result
html blast results from Primer-BLAST as 'a response

object.

taxonomy
taxonomy for the primer products from NCBI

sequence
sequence of the primer products

alignment
multiple alignment of the primer products

tree
phylogenetic tree of the reconstructed from the 'multiple alignment

See Also

primer_search, clustalo

Examples

## Not run: 
#simple search
mammals_16S = search_primer_pair(name='Mammals 16S',
 'CGGTTGGGGTGACCTCGGA', 'GCTGTTATCCCTAGGGTAACT')
#returning 1000 alignments, allow up to 3 mismatches in primer
mammals_16S = search_primer_pair(name='Mammals 16S',
 'CGGTTGGGGTGACCTCGGA', 'GCTGTTATCCCTAGGGTAACT',
 num_aligns=1000, total_primer_specificity_mismatch=3)

## End(Not run)

[Package primerTree version 1.0.3 Index]