seq_lengths {primerTree}	R Documentation

Get a summary of sequence lengths from a primerTree object

Description

Get a summary of sequence lengths from a primerTree object

Usage

seq_lengths(x, summarize = TRUE)

Arguments

x
a primerTree object.

summarize
a logical indicating if a summary should be displayed

Value

a table of sequence length frequencies

Examples

# Show the counts for each length
seq_lengths(mammals_16S)

# Plot the distribution of lengths
seqLengths <- seq_lengths(mammals_16S)
barplot(seqLengths,
 main = "Frequency of sequence lengths for 16S mammal primers",
 xlab="Amplicon length (in bp)",
 ylab=("Frequency"))

[Package primerTree version 1.0.3 Index]