summary.primerTree {primerTree}	R Documentation

Summarize a primerTree result, printing quantiles of sequence length and pairwise differences.

Description

Summarize a primerTree result, printing quantiles of sequence length and pairwise differences.

Usage

## S3 method for class 'primerTree'
summary(object, ..., probs = c(0, 0.05, 0.5, 0.95, 1),
  ranks = common_ranks)

Arguments

object
the primerTree object to summarise

...
Ignored options

probs
quantile probabilities to compute, defaults to 0, 5, 50, 95, and 100 probabilities.

ranks
ranks to show unique counts for, defaults to the common ranks

Value

invisibly returns a list containing the printed results

[Package primerTree version 1.0.3 Index]